+++ features/auth/index.ts
@@ -1,6 +1,6 @@
 export { exchangeKakaoCode, fetchNeedsOnboarding } from './auth.api';
 export { LOGIN_RETURN_COOKIE_NAME } from './auth.constants';
-export { useLoginModal } from './auth.hooks';
+export { useLoginModal } from './login-modal.hooks';
 export { openLoginModalOnPage, resolveKakaoLoginDestination, saveLoginReturnPath } from './auth.lib';
-export { KakaoLoginButton } from './components/kakao-login-button';
+export { LoginButton } from './components/login-button';
 export { SignupFlow } from './components/signup-flow';
